llvm38: version bump to 3.8.1
This bumps the version from 3.8.0 to 3.8.1.

Note that LLVM HQ forgot to update the LLVM_VERSION_PATCH variable for the configure based build system, which we use here (it was only updated for the cmake based system). The extra embedded patch fixes the version string throughout the configure script so that it is reported correctly by --version and friends.